refactor(translation): own Anthropic request normalization
